# api-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bitexroom api",
        "description": "Description of your API",
        "contact": {
            "email": "mehdints@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in a user",
                "description": "Authenticates a user and optionally generates an access token. If the user has two-factor authentication enabled, the access token is not returned, and additional verification is required.",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                },
                                                "token": {
                                                    "oneOf": [
                                                        {
                                                            "$ref": "#/components/schemas/AccessTokenResource"
                                                        }
                                                    ],
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "throttle": []
                    }
                ]
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Handles user registration, revokes previous sessions, generates a new access token, and sends a verification email.",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registration successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "$ref": "#/components/schemas/AccessTokenResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "throttle": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "The user's email address.",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The user's password.",
                        "type": "string",
                        "format": "password"
                    },
                    "captcha": {
                        "description": "Captcha response if in production.",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "email",
                    "captcha",
                    "key",
                    "password"
                ],
                "properties": {
                    "first_name": {
                        "description": "The user's first name.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "last_name": {
                        "description": "The user's last name.",
                        "type": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "description": "The user's email address.",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The user's password.",
                        "type": "string",
                        "format": "password",
                        "minLength": 8
                    },
                    "introducer_code": {
                        "description": "Referral code, if applicable.",
                        "type": "string",
                        "nullable": true
                    },
                    "key": {
                        "description": "Captcha's key required in production",
                        "type": "string"
                    },
                    "captcha": {
                        "description": "required in production.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "AccessTokenResource": {
                "properties": {
                    "access_token": {
                        "description": "The generated access token.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "name": {
                        "description": "The user's full name.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The user's email address.",
                        "type": "string",
                        "format": "email"
                    },
                    "has_two_factor": {
                        "description": "Indicates whether the user has two-factor authentication enabled.",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}